New Update Admin

+++ app/Http/Controllers/Merchant/MerchantProdukController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Merchant;
 use App\Http\Controllers\Controller;
 use Validator;
 use App\Models\Produk;
+use App\Models\Order;
 use App\Models\Pembuatan;
 use Illuminate\Http\Request;
 use Illuminate\Database\QueryException;
@@ -20,7 +21,7 @@ class MerchantProdukController extends Controller
         $userId = auth()->user()->id;
         $produks = Pembuatan::join('produks', 'pembuatans.produk_id', '=', 'produks.id')
                             ->join('users', 'produks.user_id', '=', 'users.id')
-                            ->select('pembuatans.*', 'produks.*', 'users.name')
+                            ->select('pembuatans.*', 'produks.nama_produk','produks.foto_produk','produks.harga', 'users.name')
                             ->where('produks.user_id', '=', $userId)
                             ->where('pembuatans.tanggal_jadi', '>=', now())
                             ->get();
@@ -153,8 +154,16 @@ class MerchantProdukController extends Controller
     {
         // dd($id);
         try {
-            Pembuatan::where('produk_id', $id)->delete();
-            Produk::destroy($id);
+            $order = Order::where('pembuatan_id', $id)->get();
+            // dd($order);
+            if (is_null($order)) {
+                return redirect('/penjual/product')->with('error', 'Data masih memiliki order.');
+            } else {
+                // Pembuatan::destroy($id);
+                Pembuatan::where('produk_id', $id)->delete();
+                Produk::destroy($id);
+                return redirect('/penjual/product')->with('success', 'Data berhasil dihapus.');
+            }
             
             return redirect('/penjual/product')->with('success', 'Data berhasil dihapus.');
         } catch (\Exception $e) {
